refactor: 스타일 파일 위치 변경, 패딩 추가
